Working with REST APIs - Finished

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require("express-validator");
 const path = require("path");
 const fs = require("fs");
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -19,7 +20,7 @@ exports.getPosts = (req, res, next) => {
       res.status(200).json({
         message: "Fetched posts succesfully.",
         posts: posts,
-        totalItems:  totalItems
+        totalItems: totalItems,
       });
     })
     .catch((err) => {
@@ -49,24 +50,34 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path.replace("\\", "/");
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
 
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: "Federico",
-    },
+    creator: req.userId,
   });
 
   post
     .save()
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       //Create post in db
       res.status(201).json({
         message: "Post created succesfully!",
-        post: result,
+        post: post,
+        creator: {
+          _id: creator._id,
+          name: creator.name,
+        },
       });
     })
     .catch((err) => {
@@ -82,6 +93,7 @@ exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
 
   Post.findById(postId)
+  .populate('creator') 
     .then((post) => {
       if (!post) {
         const error = new Error("Could not  find post.");
@@ -136,6 +148,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -171,12 +189,24 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
+      if(post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       // Check logged user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+     user.posts.pull(postId);
+    return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
